[JavaScript/Python | EN DateTime V2] Fixed missing resolution for years spelled out as words (#2332)
