+++ JavaScript/packages/recognizers-date-time/src/dateTime/baseDatePeriod.ts
@@ -1,7 +1,7 @@
 import { IExtractor, ExtractResult, RegExpUtility, Match, StringUtility } from "@microsoft/recognizers-text";
 import { Constants, TimeTypeConstants } from "./constants";
 import { BaseNumberExtractor, BaseNumberParser } from "@microsoft/recognizers-text-number";
-import { Token, DateTimeFormatUtil, DateTimeResolutionResult, DateUtils, DayOfWeek, StringMap } from "./utilities";
+import { Token, DateTimeFormatUtil, DateTimeResolutionResult, DateUtils, DayOfWeek, StringMap, AbstractYearExtractor } from "./utilities";
 import { BaseDurationExtractor, BaseDurationParser } from "./baseDuration";
 import { IDateTimeParser, DateTimeParseResult } from "./parsers";
 import { BaseDateExtractor, BaseDateParser } from "./baseDate";
@@ -63,7 +63,7 @@ export class BaseDatePeriodExtractor implements IDateTimeExtractor {
                 if (matchYear && matchYear.length === match.value.length) {
                     let yearStr = matchYear.groups('year').value;
                     if (StringUtility.isNullOrEmpty(yearStr)) {
-                        let year = this.getYearFromText(matchYear);
+                        let year = AbstractYearExtractor.getYearFromText(matchYear, this.config.numberParser);
                         if (!(year >= Constants.MinYearNum && year <= Constants.MaxYearNum)) {
                             addToken = false;
                         }
@@ -83,58 +83,6 @@ export class BaseDatePeriodExtractor implements IDateTimeExtractor {
             });
         });
         return tokens;
-    }
-
-    private getYearFromText(match: Match): number {
-
-        let year = -1;
-
-        let yearStr = match.groups('year').value;
-        if (!StringUtility.isNullOrEmpty(yearStr)) {
-            year = Number.parseInt(yearStr);
-            if (year < 100 && year >= Constants.MinTwoDigitYearPastNum)
-            {
-                year += 1900;
-            }
-            else if (year >= 0 && year < Constants.MaxTwoDigitYearFutureNum)
-            {
-                year += 2000;
-            }
-        }
-        else { 
-            let firstTwoYearNumStr = match.groups('firsttwoyearnum').value;
-            if (!StringUtility.isNullOrEmpty(firstTwoYearNumStr)) {
-                let er = new ExtractResult();
-                er.text = firstTwoYearNumStr;
-                er.start = match.groups('firsttwoyearnum').index;
-                er.length = match.groups('firsttwoyearnum').length;
-
-                let firstTwoYearNum = Number.parseInt(this.config.numberParser.parse(er).value);
-
-                let lastTwoYearNum = 0;
-                let lastTwoYearNumStr = match.groups('lasttwoyearnum').value;
-                if (!StringUtility.isNullOrEmpty(lastTwoYearNumStr)) {
-                    er.text = lastTwoYearNumStr;
-                    er.start = match.groups('lasttwoyearnum').index;
-                    er.length = match.groups('lasttwoyearnum').length;
-
-                    lastTwoYearNum = Number.parseInt(this.config.numberParser.parse(er).value);
-                }
-
-                if (firstTwoYearNum < 100 && lastTwoYearNum === 0 || firstTwoYearNum < 100 && firstTwoYearNum % 10 === 0 && lastTwoYearNumStr.trim().split(' ').length === 1) {
-                    year = -1;
-                }
-
-                if (firstTwoYearNum >= 100) {
-                    year = (firstTwoYearNum + lastTwoYearNum);
-                }
-                else {
-                    year = (firstTwoYearNum * 100 + lastTwoYearNum);
-                }
-            }
-        }
-
-        return year;
     }
 
     protected mergeTwoTimePoints(source: string, refDate: Date): Token[] {
@@ -283,6 +231,7 @@ export interface IDatePeriodParserConfiguration {
     dateParser: BaseDateParser
     durationExtractor: IDateTimeExtractor
     durationParser: BaseDurationParser
+    numberParser: BaseNumberParser
     monthFrontBetweenRegex: RegExp
     betweenRegex: RegExp
     monthFrontSimpleCasesRegex: RegExp
@@ -875,7 +824,7 @@ export class BaseDatePeriodParser implements IDateTimeParser {
             return result;
         }
 
-        let year = Number.parseInt(match.value, 10);
+        let year = AbstractYearExtractor.getYearFromText(match, this.config.numberParser);
         let beginDate = DateUtils.safeCreateFromValue(DateUtils.minValue(), year, 0, 1);
         let endDate = DateUtils.addDays(DateUtils.safeCreateFromValue(DateUtils.minValue(), year + 1, 0, 1), this.inclusiveEndPeriod ? -1 : 0);
         result.timex = DateTimeFormatUtil.toString(year, 4);
