+++ JavaScript/packages/recognizers-date-time/src/dateTime/chinese/datePeriodConfiguration.ts
@@ -139,6 +139,7 @@ class ChineseDatePeriodParserConfiguration implements IDatePeriodParserConfigura
     readonly dateParser: BaseDateParser
     readonly durationExtractor: ChineseDurationExtractor
     readonly durationParser: BaseDurationParser
+    readonly numberParser: BaseNumberParser
     readonly monthFrontBetweenRegex: RegExp
     readonly betweenRegex: RegExp
     readonly monthFrontSimpleCasesRegex: RegExp
